requirements.txt
# Core Dependencies
PyYAML>=6.0
python-dotenv>=1.0.0

# Async and Concurrency
asyncio-contextmanager>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.12.0

# Logging and Monitoring
structlog>=23.1.0

# Optional: For Google API integration (if using Google Search API)
# google-api-python-client>=2.0.0
# google-auth-oauthlib>=1.0.0

# Optional: For cloud deployment
# google-cloud-run>=1.0.0
# kubernetes>=25.0.0
